requirements.txt
Django==3.1.14
psycopg2-binary==2.9.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.0
